reorganised the folder structure
