+++ src/main/java/com/spartaglobal/moviesapi/data/DataLoader.java
@@ -1,5 +1,6 @@
-package com.spartaglobal.moviesapi.controller;
+package com.spartaglobal.moviesapi.data;
 
+import com.spartaglobal.moviesapi.filehandling.FileProcessor;
 import com.spartaglobal.moviesapi.model.Film;
 import com.spartaglobal.moviesapi.repository.FilmRepository;
 import java.util.List;
