+++ src/main/java/com/spartaglobal/moviesapi/filehandling/FileProcessor.java
@@ -1,9 +1,8 @@
-package com.spartaglobal.moviesapi.controller;
+package com.spartaglobal.moviesapi.filehandling;
 
 
-import com.spartaglobal.moviesapi.filehandling.CsvReader;
-import com.spartaglobal.moviesapi.filehandling.CsvWriter;
 import com.spartaglobal.moviesapi.model.Film;
+import com.spartaglobal.moviesapi.model.FilmFactory;
 import com.spartaglobal.moviesapi.validation.CsvRowValidator;
 import java.util.ArrayList;
 import java.util.List;
