+++ src/main/java/com/spartaglobal/moviesapi/model/FilmFactory.java
@@ -1,4 +1,4 @@
-package com.spartaglobal.moviesapi.controller;
+package com.spartaglobal.moviesapi.model;
 
 import com.spartaglobal.moviesapi.model.Film;
