+++ src/main/java/com/spartaglobal/moviesapi/service/FilmService.java
@@ -1,7 +1,7 @@
 package com.spartaglobal.moviesapi.service;
 
 import com.spartaglobal.moviesapi.model.Film;
-import com.spartaglobal.moviesapi.exceptions.NoFilmsInDatabaseException;
+import com.spartaglobal.moviesapi.service.exceptions.NoFilmsInDatabaseException;
 import java.util.List;
 import org.springframework.stereotype.Component;
