+++ src/main/java/com/spartaglobal/moviesapi/validation/ActorsNamesValidationRule.java
@@ -1,7 +1,7 @@
 package com.spartaglobal.moviesapi.validation;
 
-import com.spartaglobal.moviesapi.exceptions.ValidateException;
-import com.spartaglobal.moviesapi.exceptions.ValidateMoviesException.InvalidNameException;
+import com.spartaglobal.moviesapi.validation.exceptions.ValidateException;
+import com.spartaglobal.moviesapi.validation.exceptions.ValidateMoviesException.InvalidNameException;
 
 public class ActorsNamesValidationRule implements FilmValidationRule {
