+++ src/main/java/com/spartaglobal/moviesapi/validation/CsvRowValidator.java
@@ -1,6 +1,6 @@
 package com.spartaglobal.moviesapi.validation;
 
-import com.spartaglobal.moviesapi.exceptions.ValidateException;
+import com.spartaglobal.moviesapi.validation.exceptions.ValidateException;
 import java.util.ArrayList;
 import java.util.List;
