+++ src/main/java/com/spartaglobal/moviesapi/validation/GrossValidationRule.java
@@ -1,7 +1,7 @@
 package com.spartaglobal.moviesapi.validation;
 
-import com.spartaglobal.moviesapi.exceptions.ValidateException;
-import com.spartaglobal.moviesapi.exceptions.ValidateMoviesException.InvalidGross;
+import com.spartaglobal.moviesapi.validation.exceptions.ValidateException;
+import com.spartaglobal.moviesapi.validation.exceptions.ValidateMoviesException.InvalidGross;
 
 public class GrossValidationRule implements FilmValidationRule {
