+++ src/test/java/com/spartaglobal/moviesapi/controller/CsvRowValidatorTest.java
@@ -3,17 +3,17 @@ package com.spartaglobal.moviesapi.controller;
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertThrows;
 
-import com.spartaglobal.moviesapi.exceptions.ValidateMoviesException.InvalidBudgetException;
-import com.spartaglobal.moviesapi.exceptions.ValidateMoviesException.InvalidCountryException;
-import com.spartaglobal.moviesapi.exceptions.ValidateMoviesException.InvalidDurationException;
-import com.spartaglobal.moviesapi.exceptions.ValidateMoviesException.InvalidGenreException;
-import com.spartaglobal.moviesapi.exceptions.ValidateMoviesException.InvalidGross;
-import com.spartaglobal.moviesapi.exceptions.ValidateMoviesException.InvalidLanguageException;
-import com.spartaglobal.moviesapi.exceptions.ValidateMoviesException.InvalidNameException;
-import com.spartaglobal.moviesapi.exceptions.ValidateMoviesException.InvalidRatingException;
-import com.spartaglobal.moviesapi.exceptions.ValidateMoviesException.InvalidScoreException;
-import com.spartaglobal.moviesapi.exceptions.ValidateMoviesException.InvalidTitleException;
-import com.spartaglobal.moviesapi.exceptions.ValidateMoviesException.InvalidYearException;
+import com.spartaglobal.moviesapi.validation.exceptions.ValidateMoviesException.InvalidBudgetException;
+import com.spartaglobal.moviesapi.validation.exceptions.ValidateMoviesException.InvalidCountryException;
+import com.spartaglobal.moviesapi.validation.exceptions.ValidateMoviesException.InvalidDurationException;
+import com.spartaglobal.moviesapi.validation.exceptions.ValidateMoviesException.InvalidGenreException;
+import com.spartaglobal.moviesapi.validation.exceptions.ValidateMoviesException.InvalidGross;
+import com.spartaglobal.moviesapi.validation.exceptions.ValidateMoviesException.InvalidLanguageException;
+import com.spartaglobal.moviesapi.validation.exceptions.ValidateMoviesException.InvalidNameException;
+import com.spartaglobal.moviesapi.validation.exceptions.ValidateMoviesException.InvalidRatingException;
+import com.spartaglobal.moviesapi.validation.exceptions.ValidateMoviesException.InvalidScoreException;
+import com.spartaglobal.moviesapi.validation.exceptions.ValidateMoviesException.InvalidTitleException;
+import com.spartaglobal.moviesapi.validation.exceptions.ValidateMoviesException.InvalidYearException;
 import com.spartaglobal.moviesapi.validation.ActorsNamesValidationRule;
 import com.spartaglobal.moviesapi.validation.BudgetValidationRule;
 import com.spartaglobal.moviesapi.validation.CountryValidationRule;
